Make the following changes:
,

Find this element:

  <Input id="phone" type="tel" placeholder="Enter your phone number" required=true>

Relevant files:

    - https://6000-firebase-studio-1747424608398.cluster-ejd22kqny5htuv5dfowoyipt52.cloudworkstations.dev/_next/static/chunks/_353e5444._.js
    - https://6000-firebase-studio-1747424608398.cluster-ejd22kqny5htuv5dfowoyipt52.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-firebase-studio-1747424608398.cluster-ejd22kqny5htuv5dfowoyipt52.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

veeify and check and make sure the number is new and should not exist in database, add a small box in lleft of this this enter your 10 digit phone number box (non editale) and write +91 in that and make the box as big as the +91 text. this will help user to not get confused and just enter their number. if the number already exists. tell them to log in instead.